# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "EdgeCraft Swagger API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/clouds/{cloudId}/clusters/{clusterId}/backup": {
            "get": {
                "description": "클러스터의 백업 리스트",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Backup"
                ],
                "summary": "GetBackupList",
                "operationId": "GetBackupList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "클러스터의 백업 실행 (Velero)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Backup"
                ],
                "summary": "SetBackup",
                "operationId": "SetBackup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "BackResParam",
                        "name": "backresParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackResParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/clouds/{cloudId}/clusters/{clusterId}/backup/{backresId}": {
            "delete": {
                "description": "클러스터의 백업 삭제",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Backup"
                ],
                "summary": "DeleteBackup",
                "operationId": "DeleteBackup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BackRes ID",
                        "name": "backresId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
        "/clouds/{cloudId}/clusters/{clusterId}/restore": {
            "get": {
                "description": "클러스터의 복원 리스트",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Restore"
                ],
                "summary": "GetRestoreList",
                "operationId": "GetRestoreList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "클러스터의 복원 실행 (Velero)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Restore"
                ],
                "summary": "SetRestore",
                "operationId": "SetRestore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "BackResParam",
                        "name": "backresParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackResParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/clouds/{cloudId}/clusters/{clusterId}/restore/{backresId}": {
            "delete": {
                "description": "클러스터의 복원 삭제",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-Restore"
                ],
                "summary": "DeleteRestore",
                "operationId": "DeleteRestore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BackRes ID",
                        "name": "backresId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.BackResParam": {
            "type": "object",
            "properties": {
                "backresId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "isError": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}